+++ docs/release-notes.md
@@ -3,7 +3,8 @@
 * For players:
   * Improved how mod issues are listed in the console and log.
   * Revamped installer. It now...
-    * uses a new format that should be more intuitive for players;
+    * uses a new format that should be more intuitive;
+    * lets players on Linux/Mac choose the console color scheme (SMAPI will auto-detect it on Windows);
     * and validates requirements earlier.
   * Fixed custom festival maps always using spring tilesheets.
   * Fixed `player_add` command not recognising return scepter.